DF-9563 Fix errors
Former-commit-id: 8c994c0b3043945a08d978c759a8ecb3bb8c75ff

+++ Sources/Internal/Render/2D/TextBlock.cpp
@@ -699,7 +699,7 @@ void TextBlock::CalculateCacheParams()
 
         Line lineInfo;
         lineInfo.offset = 0;
-        lineInfo.length = visualText.size();
+        lineInfo.length = uint32(visualText.size());
         lineInfo.number = 0;
         lineInfo.xadvance = std::accumulate(charactersSizes.begin(), charactersSizes.end(), 0.f);
         lineInfo.visibleadvance = static_cast<float32>(textMetrics.width);
